https://github.com/d-koppenhagen/angular-tag-cloud-module/issues/31
Support for non radomization of angles when undefined in cloud data entries and added an alphabetical sort of data before sorting by weight to allow for uniform placement of words in the cloud. This can ensure that if a word cloud is updated with the same data the placement of words will not change. Previously, because only sorting by weight, there was some randomness to the word ordering when rendering.